String formatting correction.

+++ Source/Managers/Theme Manager/ThemeManager.m
@@ -40,7 +40,7 @@
 
 + (UIImage *)randomImage {
     NSInteger randomNumber = [[DataManager sharedInstance] randomNumberBetweenMin:1 andMax:6];
-    NSString *randomImageName = [NSString stringWithFormat:@"r%ld", randomNumber];
+    NSString *randomImageName = [NSString stringWithFormat:@"r%ld", (long)randomNumber];
     return [UIImage imageNamed:randomImageName] ?: [UIImage imageNamed:kDefaultRandomImageName];
 }
